updated DELETE stuff endpoint with file deletion handling

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 import { Thing } from "../database/models/Thing.js";
+import fs from "fs";
 
 export function getThings(req, res, next) {
 	Thing.find()
@@ -63,7 +64,22 @@ export function modifyThing(req, res, next) {
 }
 
 export function deleteThing(req, res, next) {
-	Thing.deleteOne({ _id: req.params.id })
-		.then(() => res.status(200).json({ message: "Objet supprimé !" }))
-		.catch(error => res.status(400).json({ error }));
+	Thing.findOne({ _id: req.params.id })
+		.then(thing => {
+			if (thing.userId != req.auth.userId) {
+				res.status(401).json({ message: "Opération non autorisée !" });
+			} else {
+				const filename = thing.imageUrl.split("/images/")[1];
+				fs.unlink(`images/${filename}`, () => {
+					Thing.deleteOne({ _id: req.params.id })
+						.then(() =>
+							res
+								.status(200)
+								.json({ message: "Objet supprimé !" })
+						)
+						.catch(error => res.status(400).json({ error }));
+				});
+			}
+		})
+		.catch(error => res.status(500).json({ error }));
 }
